requirements.txt
GitPython>=3.1.40
colorama>=0.4.6
tabulate>=0.9.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0